+++ API/RDV.js
@@ -6,41 +6,45 @@ const { FicheConsultation } = require("../Models/FicheConsultation");
 exports.ADDRDV = async (req, resp) => {
   let Cab = req.params.id;
   let name = req.body.name;
+  let id = req.user;
 
   let Name = name[0].split(" ");
   let Date = req.body.Date[0].split("T");
-  console.log(Date);
-
-  console.log(Name);
 
   const newrdv = new RDV(Date[0], Date[1], req.body.Type, "en cours");
 
   try {
-    let List = await sqlQuery(
-      ` SELECT Id FROM Patient WHERE (Nom = '${Name[0]}' && Prénom = '${Name[1]}') ||  (Nom = '${Name[1]}' && Prénom = '${Name[0]}') `
-    );
-
-    if (List.length === 0) {
-      resp.status(201).json({ message: "Ce patient n'existe pas" });
-    } else {
-      let res = await sqlQuery(
-        ` SELECT * FROM rdv WHERE Patient = '${List[0].Id}' and Etat = 'en cours' `
+    if (id !== Cab)
+      resp
+        .status(201)
+        .json({ message: "Vous ne pouvez effectuer cette operation" });
+    else {
+      let List = await sqlQuery(
+        ` SELECT Id FROM Patient WHERE (Nom = '${Name[0]}' && Prénom = '${Name[1]}') ||  (Nom = '${Name[1]}' && Prénom = '${Name[0]}') `
       );
 
-      console.log(res);
+      if (List.length === 0) {
+        resp.status(201).json({ message: "Ce patient n'existe pas" });
+      } else {
+        let res = await sqlQuery(
+          ` SELECT * FROM rdv WHERE Patient = '${List[0].Id}' and Etat = 'en cours' `
+        );
 
-      if (res.length !== 0)
-        resp
-          .status(201)
-          .json({ message: "Ce patient a deja un rendez vous en cours" });
-      else {
-        newrdv.Cabinet = Cab;
-        newrdv.Patient = List[0].Id;
+        console.log(res);
 
-        let query = `INSERT INTO rdv Set ?`;
+        if (res.length !== 0)
+          resp
+            .status(201)
+            .json({ message: "Ce patient a deja un rendez vous en cours" });
+        else {
+          newrdv.Cabinet = Cab;
+          newrdv.Patient = List[0].Id;
 
-        if (sqlQuery(query, newrdv)) {
-          resp.status(201).json({ message: "Rdv added succesfully" });
+          let query = `INSERT INTO rdv Set ?`;
+
+          if (sqlQuery(query, newrdv)) {
+            resp.status(201).json({ message: "Rdv added succesfully" });
+          }
         }
       }
     }
@@ -51,15 +55,22 @@ exports.ADDRDV = async (req, resp) => {
 
 exports.RDVList = async (req, resp) => {
   let Cab = req.params.id;
+  let id = req.user;
 
   try {
-    let List = await sqlQuery(
-      `SELECT * FROM Patient join rdv on Patient.Id = rdv.Patient WHERE Cabinet = '${Cab}'`
-    );
+    if (id !== Cab)
+      resp
+        .status(201)
+        .json({ message: "Vous ne pouvez effectuer cette operation" });
+    else {
+      let List = await sqlQuery(
+        `SELECT * FROM Patient join rdv on Patient.Id = rdv.Patient WHERE Cabinet = '${Cab}'`
+      );
 
-    resp.status(201).json({
-      listCons: List,
-    });
+      resp.status(201).json({
+        listCons: List,
+      });
+    }
   } catch (err) {
     console.log(err.message);
   }
@@ -67,10 +78,18 @@ exports.RDVList = async (req, resp) => {
 
 exports.DeleteRDV = async (req, resp) => {
   let RDV = req.params.id;
+  let id = req.user;
+  let Cab = req.ID;
 
   try {
-    if (sqlQuery(`DELETE FROM rdv WHERE Id = '${RDV}'`))
-      resp.status(201).json({ message: "RDV supprimer avec success" });
+    if (id !== Cab)
+      resp
+        .status(201)
+        .json({ message: "Vous ne pouvez effectuer cette operation" });
+    else {
+      if (sqlQuery(`DELETE FROM rdv WHERE Id = '${RDV}'`))
+        resp.status(201).json({ message: "RDV supprimer avec success" });
+    }
   } catch {
     (err) => console.log(err.message);
   }
@@ -78,14 +97,22 @@ exports.DeleteRDV = async (req, resp) => {
 
 exports.UpdateRDV = async (req, resp) => {
   let RDV = req.params.id;
+  let id = req.user;
+  let Cab = req.ID;
 
   let Date = req.body.Date[0].split("T");
 
   try {
-    let query1 = `UPDATE rdv Set Date = '${Date[0]}', Heure = '${Date[1]}', Type = '${req.body.Type}' WHERE id = '${RDV}'`;
+    if (id !== Cab)
+      resp
+        .status(201)
+        .json({ message: "Vous ne pouvez effectuer cette operation" });
+    else {
+      let query1 = `UPDATE rdv Set Date = '${Date[0]}', Heure = '${Date[1]}', Type = '${req.body.Type}' WHERE id = '${RDV}'`;
 
-    if (sqlQuery(query1)) {
-      resp.status(201).json({ message: "Update Success" });
+      if (sqlQuery(query1)) {
+        resp.status(201).json({ message: "Update Success" });
+      }
     }
   } catch {
     (err) => console.log(err.message);
@@ -94,16 +121,22 @@ exports.UpdateRDV = async (req, resp) => {
 
 exports.RDVNbr = async (req, resp) => {
   let Cab = req.params.id;
-  console.log(Cab);
+  let id = req.user;
 
   try {
-    let ConsList = await sqlQuery(
-      `SELECT COUNT(*) as nbr FROM rdv WHERE rdv.Cabinet = '${Cab}' `
-    );
+    if (id !== Cab)
+      resp
+        .status(201)
+        .json({ message: "Vous ne pouvez effectuer cette operation" });
+    else {
+      let ConsList = await sqlQuery(
+        `SELECT COUNT(*) as nbr FROM rdv WHERE rdv.Cabinet = '${Cab}' `
+      );
 
-    resp.status(201).json({
-      NbrRDV: ConsList[0].nbr,
-    });
+      resp.status(201).json({
+        NbrRDV: ConsList[0].nbr,
+      });
+    }
   } catch (err) {
     console.log(err.message);
   }
@@ -112,46 +145,55 @@ exports.RDVNbr = async (req, resp) => {
 exports.TypeUpdate = async (req, resp) => {
   let RDV = req.params.id;
   let Patient = req.params.Patient;
+  let id = req.user;
+  let Cab = req.ID;
 
   try {
-    let consult = await sqlQuery(
-      `SELECT * FROM consultation WHERE RDV = '${RDV}'`
-    );
-    console.log(consult);
-    if (consult.length !== 0) {
-      resp.status(201).json({
-        message: "Il y a deja une consultation qui correspond a ce rendez-vous",
-      });
-    } else {
-      let result = await sqlQuery(
-        `UPDATE rdv SET  Etat = 'Traité' WHERE Id = '${RDV}'`
+    if (id !== Cab)
+      resp
+        .status(201)
+        .json({ message: "Vous ne pouvez effectuer cette operation" });
+    else {
+      let consult = await sqlQuery(
+        `SELECT * FROM consultation WHERE RDV = '${RDV}'`
       );
-
-      let newconsultation = new Consultation("", "", 0, "", RDV);
-
-      let newFichConsultation = new FicheConsultation(
-        "",
-        "",
-        "",
-        "",
-        "",
-        new Date(Date.now())
-      );
-
-      let query = `INSERT INTO consultation Set ?`;
-      let query1 = `INSERT INTO fiche_consultation Set ?`;
-
-      if (sqlQuery(query, newconsultation)) {
-        let res = await sqlQuery(`SELECT * FROM consultation`);
-
-        let dossier = await sqlQuery(
-          `SELECT Id FROM dossier_medical WHERE Patient = '${Patient}'`
+      console.log(consult);
+      if (consult.length !== 0) {
+        resp.status(201).json({
+          message:
+            "Il y a deja une consultation qui correspond a ce rendez-vous",
+        });
+      } else {
+        let result = await sqlQuery(
+          `UPDATE rdv SET  Etat = 'Traité' WHERE Id = '${RDV}'`
         );
 
-        newFichConsultation.Consultation = res[res.length - 1].Id;
-        newFichConsultation.Dossier_medical = dossier[0].Id;
-        if (sqlQuery(query1, newFichConsultation)) {
-          resp.status(201).json({ message: "Traitement validé" });
+        let newconsultation = new Consultation("", "", 0, "", RDV);
+
+        let newFichConsultation = new FicheConsultation(
+          "",
+          "",
+          "",
+          "",
+          "",
+          new Date(Date.now())
+        );
+
+        let query = `INSERT INTO consultation Set ?`;
+        let query1 = `INSERT INTO fiche_consultation Set ?`;
+
+        if (sqlQuery(query, newconsultation)) {
+          let res = await sqlQuery(`SELECT * FROM consultation`);
+
+          let dossier = await sqlQuery(
+            `SELECT Id FROM dossier_medical WHERE Patient = '${Patient}'`
+          );
+
+          newFichConsultation.Consultation = res[res.length - 1].Id;
+          newFichConsultation.Dossier_medical = dossier[0].Id;
+          if (sqlQuery(query1, newFichConsultation)) {
+            resp.status(201).json({ message: "Traitement validé" });
+          }
         }
       }
     }
